+++ app/src/main/java/com/nickoehler/brawlhalla/ranking/domain/StatLegend.kt
@@ -26,5 +26,5 @@ data class StatLegend(
     val timeHeldWeaponTwo: Int,
     val xp: Int,
     val level: Int,
-    val xpPercentage: Float,
+    val xpPercentage: Double,
 )
